Fix premium badge alignment in tile grid
