+++ ui/scss/component/section.scss
@@ -44,7 +44,7 @@
 .section__header--actions {
   padding-top: var(--spacing-m);
   margin-top: var(--spacing-m);
-  margin-bottom: var(--spacing-m);
+  margin-bottom: var(--spacing-l);
   display: flex;
   align-items: flex-start;
   justify-content: space-between;
