+++ ui/scss/themes/dark.scss
@@ -160,8 +160,8 @@
   --color-scrollbar-track-bg: transparent;
   --color-body-scrollbar-track-bg: rgba(0, 0, 0, 0.5);
 
-  --background-shade: linear-gradient(-180deg, rgba(25, 25, 25, 0.2) 2%, #202020),
-    radial-gradient(circle at 50% 117%, rgba(25, 25, 25, 0.2) 0, #202020 100%);
+  --background-shade: linear-gradient(-180deg, rgba(25, 25, 25, 0.3) 2%, #202020),
+    radial-gradient(circle at 50% 117%, rgba(25, 25, 25, 0.4) 0, #202020 100%);
 
   --mui-background: #000;
 }
